feat: move to messages/chainlock
