+++ index.js
@@ -49,7 +49,7 @@ bitcore.errors = require('./lib/errors');
 // main bitcoin library
 bitcore.Address = require('./lib/address');
 bitcore.Block = require('./lib/block');
-bitcore.ChainLock = require('./lib/chainlock/chainlock');
+bitcore.ChainLock = require('./lib/messages/chainlock');
 bitcore.MerkleBlock = require('./lib/block/merkleblock');
 bitcore.QuorumEntry = require('./lib/deterministicmnlist/QuorumEntry');
 bitcore.SimplifiedMNList = require('./lib/deterministicmnlist/SimplifiedMNList');
