+++ lib/messages/chainlock.js
