+++ typings/messages/ChainLock.d.ts
